pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "A simple task management system built with Django & Django REST Framework (DRF) using Test-Driven Development (TDD)."
authors = [
    {name = "Marcos Amorim"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"
pre-commit = "^4.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_manager.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=tasks --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py39']
exclude = '''
/(
    migrations
  | __pycache__
  | venv
  | env
)/
'''
